Reorganize code for speciality checker
